Don't return seperate collection and collectionTree fields in API
