+++ common/display/src/main/scala/uk/ac/wellcome/display/models/v2/DisplayWorkV2.scala
@@ -5,7 +5,6 @@ import io.swagger.v3.oas.annotations.media.Schema
 
 import uk.ac.wellcome.display.models._
 import uk.ac.wellcome.models.work.internal.IdentifiedWork
-import uk.ac.wellcome.display.models.DisplayCollectionTree
 
 @Schema(
   name = "Work",
@@ -95,10 +94,6 @@ case class DisplayWorkV2(
     `type` = "Collection",
     description = "The collection a work is part of."
   ) collection: Option[DisplayCollection] = None,
-  @Schema(
-    `type` = "CollectionTree",
-    description = "The partially expanded collection tree for this work."
-  ) collectionTree: Option[DisplayCollectionTree] = None,
   @JsonKey("type") @Schema(name = "type") ontologyType: String = "Work"
 ) extends DisplayWork
 
@@ -151,11 +146,7 @@ case object DisplayWorkV2 {
         if (includes.notes)
           Some(DisplayNote.merge(work.data.notes.map(DisplayNote(_))))
         else None,
-      duration = work.data.duration,
-      collection =
-        if (includes.collection)
-          work.data.collection.map(DisplayCollection(_))
-        else None
+      duration = work.data.duration
     )
 
   def apply(work: IdentifiedWork): DisplayWorkV2 =
